Animate background component
Note: I still have an issue where there can be multiple concurrent animations - there should only ever be one at a time. I might need to generate an ID per animation, and use that in the Update method to determine if it's the right one (instead of using the viewport size).